Store current environment in redux
