+++ src/actions/projects.actions.js
@@ -1,6 +1,17 @@
-export function selectProject(project) {
-    return {
-        type: 'SELECT_PROJECT',
-        data: project
-    };
-}
+export const projectActions = {
+    SELECT_PROJECT: 'SELECT_PROJECT',
+    selectProject(project) {
+        return {
+            type: projectActions.SELECT_PROJECT,
+            data: project
+        };
+    },
+
+    SET_ENVIRONMENT: 'SET_ENVIRONMENT',
+    setEnvironment(environment) {
+       return {
+            type: projectActions.SET_ENVIRONMENT,
+            data: environment
+       };
+    }
+};
