+++ src/components/networkAccountSelector/accountsList.js
@@ -1,0 +1,106 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
+import Tooltip from '../tooltip';
+import style from './style.less';
+import { IconTrash, IconEdit } from '../icons';
+
+export class AccountsList extends Component {
+    render() {
+        var accounts, chosenAccount;
+        const project = this.props.router.control.getActiveProject();
+        if (!project) {
+            // Setup default account just for show.
+            accounts = [
+                {
+                    getName: () => {
+                        return 'Default';
+                    },
+                },
+            ];
+            chosenAccount = 'Default';
+        } else {
+            chosenAccount = project.getAccount();
+            const accountsItem = project.getHiddenItem('accounts');
+            accounts = accountsItem.getChildren();
+        }
+
+        const renderedAccounts = accounts.map((account, index) => {
+            const cls = {};
+            cls[style.accountLink] = true;
+            if (account.getName() == chosenAccount)
+                cls[style.accountLinkChosen] = true;
+
+            var deleteButton;
+            if (index !== 0) {
+                deleteButton = (
+                    <button
+                        className="btnNoBg"
+                        onClick={e => {
+                            this.props.onAccountDelete(e, index);
+                        }}
+                    >
+                        <Tooltip title="Delete">
+                            <IconTrash />
+                        </Tooltip>
+                    </button>
+                );
+            } else {
+                deleteButton = (
+                    <button className="btnNoBg">
+                        <i>&nbsp;&nbsp;&nbsp;&nbsp;</i>
+                    </button>
+                );
+            }
+
+            return (
+                <div key={index}>
+                    <div
+                        className={classnames(cls)}
+                        onClick={e => {
+                            e.preventDefault();
+                            this.props.onAccountChosen(account.getName());
+                        }}
+                    >
+                        <div>{account.getName()}</div>
+                        <div style={{marginLeft: 'auto'}}>
+                            <button
+                                className="btnNoBg"
+                                onClick={e => {
+                                    if (this.props.functions.EVM.isReady()) {
+                                        this.props.onAccountEdit(e, index);
+                                    } else {
+                                        console.log("EVM is not ready!");
+                                    }
+                                }}
+                            >
+                                <Tooltip title="Edit Account">
+                                    <IconEdit />
+                                </Tooltip>
+                            </button>
+                            {deleteButton}
+                        </div>
+                    </div>
+                </div>
+            );
+        });
+        return (
+            <div className={classnames([style.accounts])}>
+                <div className={style.title}>Select an Account</div>
+                {renderedAccounts}
+                <div className={style.newAccount} onClick={this.props.onNewAccountClicked}>
+                    <button className="btnNoBg">
+                        + New Account
+                    </button>
+                </div>
+            </div>
+        );
+    }
+}
+
+AccountsList.propTypes = {
+    onAccountChosen: PropTypes.func.isRequired,
+    onAccountEdit: PropTypes.func.isRequired,
+    onAccountDelete: PropTypes.func.isRequired,
+    onNewAccountClicked: PropTypes.func.isRequired,
+};
