+++ src/components/projecteditor/ProjectEditor.js
@@ -116,22 +116,8 @@ export default class ProjectEditor extends Component {
     };
 
     render() {
-        let endpoint = '';
-        let project;
-        if (this.props.router && this.props.router.control) {
-            project =
-                this.props.router.control &&
-                this.props.router.control.getActiveProject();
-            if (project) {
-                const network = project.getEnvironment();
-                endpoint = (
-                    this.props.functions.networks.endpoints[network] || {}
-                ).endpoint;
-            }
-        }
-
         const { displayTransactionsPanel, previewSidePanel, toggleTransactionsHistoryPanel,
-                previewSidePanelActions } = this.props;
+                previewSidePanelActions, selectedEnvironment } = this.props;
 
         return (
             <div className={style.projecteditor}>
@@ -170,6 +156,7 @@ export default class ProjectEditor extends Component {
                                         dragging={this.state.sidePanelDragging}
                                         router={this.props.router}
                                         onClose={toggleTransactionsHistoryPanel}
+                                        selectedEnvironment={selectedEnvironment.name}
                                     /> }
 
                                     { previewSidePanel.open && 
@@ -177,11 +164,12 @@ export default class ProjectEditor extends Component {
                                         dragging={this.state.sidePanelDragging}
                                         {...previewSidePanel}
                                         {...previewSidePanelActions}
+                                        selectedEnvironment={selectedEnvironment.name}
                                     /> }
                                     
                                 </SplitterLayout>
                             
-                                <BottomBar endpoint={endpoint} />
+                                <BottomBar endpoint={selectedEnvironment.endpoint} />
                             </div>
                         </SplitterLayout>
                     </div>
