+++ src/components/projecteditor/index.js
@@ -16,11 +16,13 @@
 
 import { connect } from 'react-redux';
 import ProjectEditor from './ProjectEditor';
+import { getSelectedEnvironment } from '../../selectors/projects';
 import { sidePanelsActions } from '../../actions';
 
 const mapStateToProps = state => ({
     displayTransactionsPanel: state.sidePanels.showTransactionsHistory,
-    previewSidePanel: state.sidePanels.preview
+    previewSidePanel: state.sidePanels.preview,
+    selectedEnvironment: getSelectedEnvironment(state)
 });
 
 const mapDispatchToProps = (dispatch) => {
