+++ src/epics/index.js
@@ -1,7 +1,9 @@
 import { settingsEpics } from './settings';
 import { sidePanelsEpics } from './sidePanels';
+import { projectsEpics } from './projects';
 
 export const epics = [
     ...settingsEpics,
-    ...sidePanelsEpics
+    ...sidePanelsEpics,
+    ...projectsEpics
 ];
