+++ src/epics/projects/index.js
@@ -1,0 +1,5 @@
+import { environmentUpdateEpic } from './environmentUpdate.epic';
+
+export const projectsEpics = [
+    environmentUpdateEpic
+];
