+++ src/selectors/projects.js
@@ -1,3 +1,3 @@
 export const getSelectedProject = state => state.projects.selectedProject;
 export const getSelectedProjectId = state => state.projects.selectedProject.id;
-
+export const getSelectedEnvironment = state => state.projects.selectedProject.selectedEnvironment;
