+++ src/services/preview.service.js
@@ -6,6 +6,7 @@ let projectItem = null;
 let exportableDappHtml = null;
 let iframeId = null;
 let disableAccounts = false;
+let environment = null;
 
 export const previewService = {
     superProvider: null,
@@ -13,7 +14,7 @@ export const previewService = {
     init(wallet) {
         window.addEventListener('message', async (e) => {
             if (e.data.type === 'window-ready' && this.projectItem) {
-                const builtProject = await buildProjectHtml(this.projectItem, wallet, this.disableAccounts);
+                const builtProject = await buildProjectHtml(this.projectItem, wallet, this.disableAccounts, environment);
                 exportableDappHtml = builtProject.exportableContent;
                 if (e.source) {
                     e.source.postMessage({ type: 'set-content', payload: builtProject.content }, '*');
@@ -32,13 +33,17 @@ export const previewService = {
                 context: 'Contract interaction',
                 network
             });
-        });
+        }, () => environment.name);
     },
 
     get projectItem() { return projectItem; },
     set projectItem(value) { 
         projectItem = value;
         exportableDappHtml = null;
+    },
+
+    setEnvironment(value) {
+        environment = value
     },
 
     get disableAccounts() { return disableAccounts; },
