+++ src/services/utils/buildProjectHtml.js
@@ -1,5 +1,3 @@
-import Networks from '../../networks';
-
 function getProviderHTML(endpoint, accounts) {
     const js =
         `<script type="text/javascript" src="${window.location.origin}/static/js/web3provider.js?ts=${Date.now()}"></script>
@@ -74,7 +72,7 @@ function makeFileName(path, network, suffix) {
     return `/build${dir}${contractName}/${contractName}.${network}.${suffix}`;
 }
 
-function getAccountAddress(project, wallet, disableAccounts) {
+function getAccountAddress(project, wallet, disableAccounts, env) {
     if (disableAccounts) { return null; }
 
     // Check given account, try to open and get address, else return [].
@@ -82,7 +80,6 @@ function getAccountAddress(project, wallet, disableAccounts) {
     if (accountName === '(locked)') { return []; }
     if (!accountName) { return []; }
 
-    const env = project.getEnvironment();
     const accounts = project.getHiddenItem('accounts');
     const account = accounts.getByName(accountName);
     const accountIndex = account.getAccountIndex(env);
@@ -120,7 +117,7 @@ function getAccountAddress(project, wallet, disableAccounts) {
  * @param {*} wallet 
  * @param {*} disableAccounts
  */
-export function buildProjectHtml(project, wallet, disableAccounts) {
+export function buildProjectHtml(project, wallet, disableAccounts, environment) {
     return new Promise((resolve, reject) => {
 
         let js, css, html, contractjs;
@@ -153,8 +150,6 @@ export function buildProjectHtml(project, wallet, disableAccounts) {
                                     .map(contract => contract.getName());
 
                                 const jsfiles = [];
-                                const network = project.getEnvironment();
-                                const endpoint = (Networks[network] || {}).endpoint;
 
                                 for (const contractName of contracts) {
                                     const contract = project
@@ -162,7 +157,7 @@ export function buildProjectHtml(project, wallet, disableAccounts) {
                                         .getByName(contractName);
 
                                     const src = contract.getSource();
-                                    const jssrc = makeFileName(src, network, "js");
+                                    const jssrc = makeFileName(src, environment.name, "js");
                                     jsfiles.push(jssrc);
                                 }
 
@@ -178,8 +173,8 @@ export function buildProjectHtml(project, wallet, disableAccounts) {
                                         css,
                                         contractjs + '\n' + js,
                                         project.getTitle(),
-                                        endpoint,
-                                        getAccountAddress(project, wallet, disableAccounts)
+                                        environment.endpoint,
+                                        getAccountAddress(project, wallet, disableAccounts, environment.name)
                                     );
 
                                     // exportable version.
