--- src/store/projects.js
@@ -1,6 +1,0 @@
-export default {
-    selectedProject: {
-        id: 0,
-        name: ''
-    }
-};
